# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "Api for all the function",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/recipe-categories": {
            "get": {
                "tags": [
                    "Recipe Categories"
                ],
                "summary": "Get all recipe categories",
                "description": "Returns a list of all recipe categories with associated recipes and categories.",
                "operationId": "getRecipeCategoriesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "Recipe": {
                                                "type": "string",
                                                "example": "Recipe Name"
                                            },
                                            "Category": {
                                                "type": "string",
                                                "example": "Category Name"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Recipe Categories"
                ],
                "summary": "Store new recipe category associations",
                "description": "Creates new associations between recipes and a category in the database.",
                "operationId": "storeRecipeCategory",
                "requestBody": {
                    "description": "RecipeCategory object containing recipe IDs and category ID",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "recipe_id",
                                    "category_id"
                                ],
                                "properties": {
                                    "recipe_id": {
                                        "description": "Array of recipe IDs",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully categorize recipe(s)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully categorize recipe."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    }
}